% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityIntegral.R
\name{densityIntegral}
\alias{densityIntegral}
\title{Compute the time integral of the diffusion kernel at a single trap location.}
\usage{
densityIntegral(startTime, endTime, trapLocation, releaseXY, releaseTimes,
  releaseNumbers, muX, muY, sigma, deathRate)
}
\arguments{
\item{startTime}{is a numeric value and is the time at which to start integrating the diffusion kernel.}

\item{endTime}{is a numeric value and is the time at which to stop integrating the diffusion kernel.}

\item{trapLocation}{is a matrix or vector containing two elements, the first containing the trap's Easting and the second containing the trap's Northing.}

\item{releaseXY}{is a matrix with two columns containing the locations of relevant releases.  The first column should contain the Eastings of the release locations and the second column should contain Northings.}

\item{releaseTimes}{is a vector containing the time (in Days) into the experiment that the releases occurred.  There should be one element of releaseTimes for each row of releaseXY.}

\item{releaseNumbers}{is a vector containing the number of individuals released at each release location in releaseXY.  There should be one element of releaseNumbers for each row of releaseXY.}

\item{muX}{is a numeric value that corresponds to the drift in the orientation of the east-west axis that should be applied to the diffusion kernel over time.  muX should be in units of distance per day, where distance is in units of Easting and Northings used to specify trap locations and released locations.}

\item{muY}{is a numeric value that corresponds to the drift in the orientation of the north-south axis that should be applied to the diffusion kernel over time.  muY should be in units of distance per day, where distance is in units of Easting and Northings used to specify trap locations and released locations.}

\item{sigma}{is a numeric value that corresponds to the standard deviation of the diffusion kernel in the east-west and north-south directions at one day post release.}

\item{deathRate}{is a numeric value that corresponds to the decay parameter of an exponential decay model.  This model is used to model the death of individuals over time.  The percent daily mortality rate can be calculated as 100*(1 - exp(-lambda))}
}
\value{
the function returns a single numeric value equal to the integrals of the diffusion kernels (one from each release location) at the trap location over the time interval [startTime, endTime].
}
\description{
\code{densityIntegral} approximates the time integral (via quadrature) of a gaussian pdf that is evolving via a process of diffusion. We model the time-varying standard deviation of the gaussian pdf as sqrt(t)*sigma in both the x and y directions.  We also allow for a time-homogeneous drift in the mean which is governed by parameters muX and muY.  We also allow for the dispersing particles (e.g. insects) to die over time according to a time-homogeneous death rate (deathRate).  The integral is computed over the interval [startTime, endTime] at locations in the rows of the trapLocations matrix.  We assume that particles were released from the locations specified in the rows of the matrix releaseLocations with the numbers released and the times of release specified in the vectors releaseNumbers and releaseTimes respectively (one element for each row of releaseLocations).
}
